+++ src/lib.rs
@@ -1,10 +1,10 @@
 mod changes;
 mod elements;
-mod graph;
+mod graph_view;
 mod settings;
 mod state;
 
 pub use self::changes::{Changes, ChangesEdge, ChangesNode};
 pub use self::elements::{Edge, Elements, Node};
-pub use self::graph::GraphView;
+pub use self::graph_view::GraphView;
 pub use self::settings::Settings;
